moved test_query_postgres_temp_db
